+++ src/bilder/components/hashicorp/nomad/models.py
@@ -1,4 +1,4 @@
-from ol_configuration_management.lib.model_helpers import OLBaseSettings
+from bilder.lib.model_helpers import OLBaseSettings
 
 
 class NomadConfig(OLBaseSettings):
