+++ src/bilder/lib/linux_helpers.py
