+++ src/bilder/lib/model_helpers.py
